make flyway config, and create dynamic migrations after school creation

+++ backend/school-service/src/main/java/org/schoolservice/dto/request/SchoolValidationDTO.java
@@ -17,6 +17,6 @@ public record SchoolValidationDTO(
         int professionalTax,
         int cnssAffiliation,
         int commercialRegister,
-        int commonBusinessIdentifier
+        Long commonBusinessIdentifier
 ) {
 }
